% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_dc.R
\name{vector_dc}
\alias{vector_dc}
\alias{numeric_dc}
\alias{integer_dc}
\title{Allocate vectors in .C64()}
\usage{
vector_dc(mode = "logical", length = 0L)

numeric_dc(length = 0)

integer_dc(length = 0)
}
\arguments{
\item{mode}{character vector of length 1. Storage mode of the vector.}

\item{length}{numeric vector of length 1. Length of the vector.}
}
\value{
object of class \code{vector_dc} and \code{list}.
}
\description{
\code{vector_dc} and its shortcuts \code{numeric_dc} and
\code{integer_dc} are helper functions used in calls to \code{\link{.C64}}.
They return an R object of class \code{c("vector_dc", "list")}, 
which contains information on the type and length of the vector to allocate.
Using \code{vector_dc} together with \code{INTENT = "w"} argument of \code{\link{.C64}}
leads to performance gains by avoiding unnecessary castings and copies.
}
\examples{
vector_dc("integer", 20)
}
